BDO-782: Changed file size type all over to accommodate large values.

+++ app/connectors/model/ContentLengthRange.scala
@@ -16,4 +16,4 @@
 
 package connectors.model
 
-case class ContentLengthRange(min: Int, max: Int)
+case class ContentLengthRange(min: Long, max: Long)
